remove footgun from test code
The sample editor component in the tests works fine
in isolation, but when used as sample starter code
for a real application it can wreak havoc, since it
removes all listeners on the handle on teardown.
This changes the behavior to only remove the listener
used by this editor component.