remove changeAt indirection
Context: when this plugin was originally written automerge did not always perform well enough to be sure that we could update the document on every keystroke. In order to allow batching updates to the document the plugin was designed to apply codemirror transactions to the editor state immediately and then later reconcile these transactions with automerge. The reconciliation step worked by queueing up the codemirror transactions, reversing them on the editor state, then converting them into an automerge change and then finally converting the automerge diff into an equivalent transaction to apply to the editor state.

Problem: this architecture meant that codemirror plugins which interact with the codemirror history (such as redo) don't work because the transactions they are expecting to work with are reversed.

Solution: now that automerge is fast enough to run on every keystroke we do away with the transaction queue and just directly apply the transaction to the editor state and then also apply the transaction to the automerge document. When we receive changes over the network we convert the corresponding automerge diff to codemirror changes and apply them directly. This means that codemirror plugins (such as the history plugin) work without modification.